requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
aioxrpy==1.0.0
arrow==0.17.0
asgiref==3.3.1
asn1crypto==1.4.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
base32-lib==1.0.2
base58==2.0.0
bit==0.7.2
bitcash==0.6.0
bitcoinlib==0.5.2
bitsv==0.11.3
boto3==1.15.6
botocore==1.18.6
cashaddress-regtest==1.1.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==6.6
cmc==2.0.3.1
coincurve==13.0.0
colorama==0.4.3
cryptography==3.1.1
Cython==0.29.21
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.0
Django==2.2.6
django-allauth==0.42.0
django-background-tasks==1.2.5
django-braces==1.14.0
django-compat==1.0.15
django-cors-headers==3.4.0
django-easy-select2==1.5.7
django-oauth-toolkit==1.3.2
django-otp==1.0.0
django-phone-field==1.8.1
django-storages==1.10.1
django-with-extra-context-admin==0.1.0
djangorestframework==3.11.0
docker-py==1.8.0
drf-social-oauth2==1.0.8
ecdsa==0.16.0
eth-hash==0.2.0
eth-typing==2.2.2
flake8==3.8.4
funcsigs==1.0.2
future==0.18.2
gunicorn==20.0.4
hexbytes==0.2.1
idna==2.9
importlib-metadata==2.0.0
iniconfig==1.1.1
jmespath==0.10.0
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
mccabe==0.6.1
mnemonic==0.13
multidict==5.0.2
mysqlclient==1.4.6
numpy==1.19.4
oauthlib==3.1.0
onetimepass==1.0.1
packaging==20.4
pandas==1.1.4
path==15.0.0
path.py==12.5.0
pbkdf2==1.3
pbr==5.4.5
pexpect==4.8.0
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.0.0a3
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
pyaes==1.6.1
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyotp==2.4.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-bitcoinrpc==1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
qrcode==6.1
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scrypt==0.8.17
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.20
sqlparse==0.3.0
stevedore==1.32.0
toml==0.10.2
toolz==0.11.1
tqdm==4.54.0
twilio==6.50.1
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==16.7.5
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
websocket-client==0.57.0
whatsonchain==0.0.3
whitenoise==5.2.0
yarl==1.6.3
zipp==3.4.0
